Reduced available threads and worker construction
Reduced the default number of threads from 6 to 4.

Workers are only constructed if there are less workers waiting for an
open thread than there are threads (i.e. if there are 4 threads and 4
workers waiting, a new worker will not be constructed). This prevents
the number of workers from piling up and filling up the RAM, which
causes the program to slow down.